Add files via upload
We have been provided with data regarding US elections, general county information, life expectancies, economic data and more.
This assignment will be focused on the elections to the House of Representatives that occur every other year. Furthermore, rather than attempting to estimate how each congressional district within a state voted, we will be interested in the overall number of votes per state (voter turnout) and who had more votes in a particular state per election, a Democrat, or a Republican.